docs: 댓글 API 문서 작성

+++ src/main/java/com/project/zipmin/controller/CommentController.java
@@ -12,13 +12,13 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.project.zipmin.api.ApiException;
 import com.project.zipmin.api.ApiResponse;
+import com.project.zipmin.api.CommentErrorCode;
 import com.project.zipmin.api.CommentSuccessCode;
 import com.project.zipmin.api.VoteErrorCode;
 import com.project.zipmin.dto.CommentCreateRequestDto;
@@ -29,8 +29,27 @@ import com.project.zipmin.dto.CommentUpdateRequestDto;
 import com.project.zipmin.dto.CommentUpdateResponseDto;
 import com.project.zipmin.service.CommentService;
 import com.project.zipmin.service.UserService;
+import com.project.zipmin.swagger.CommentCreateFailResponse;
+import com.project.zipmin.swagger.CommentCreateSuccessResponse;
+import com.project.zipmin.swagger.CommentDeleteFailResponse;
+import com.project.zipmin.swagger.CommentDeleteSuccessResponse;
+import com.project.zipmin.swagger.CommentForbiddenResponse;
+import com.project.zipmin.swagger.CommentInvalidInputResponse;
+import com.project.zipmin.swagger.CommentNotFoundResponse;
+import com.project.zipmin.swagger.CommentUnauthorizedAccessResponse;
+import com.project.zipmin.swagger.CommentUpdateFailResponse;
+import com.project.zipmin.swagger.CommentUpdateSuccessResponse;
+import com.project.zipmin.swagger.InternalServerErrorResponse;
+
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.Parameter;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
+import io.swagger.v3.oas.annotations.tags.Tag;
 
 @RestController
+@Tag(name = "Comment API", description = "댓글 관련 API")
 public class CommentController {
 	
 	@Autowired
@@ -58,26 +77,76 @@ public class CommentController {
 			commentPage = commentService.getCommentPageByTablenameAndRecodenumOrderByLikecount(tablename, recodenum, pageable);
 		}
 		
-		return ResponseEntity.status(CommentSuccessCode.COMMENT_READ_SUCCESS.getStatus())
-        		.body(ApiResponse.success(CommentSuccessCode.COMMENT_READ_SUCCESS, commentPage));
+		return ResponseEntity.status(CommentSuccessCode.COMMENT_READ_LIST_SUCCESS.getStatus())
+        		.body(ApiResponse.success(CommentSuccessCode.COMMENT_READ_LIST_SUCCESS, commentPage));
 	}
 	
 	
 	
 	// 댓글 작성
+	@Operation(
+	    summary = "댓글 작성",
+	    description = "댓글을 작성합니다."
+	)
+	@ApiResponses(value = {
+		@io.swagger.v3.oas.annotations.responses.ApiResponse(
+				responseCode = "201",
+				description = "댓글 작성 성공",
+				content = @Content(
+						mediaType = "application/json",
+						schema = @Schema(implementation = CommentCreateSuccessResponse.class))),
+		@io.swagger.v3.oas.annotations.responses.ApiResponse(
+				responseCode = "400",
+				description = "댓글 작성 실패",
+				content = @Content(
+						mediaType = "application/json",
+						schema = @Schema(implementation = CommentCreateFailResponse.class))),
+		@io.swagger.v3.oas.annotations.responses.ApiResponse(
+				responseCode = "400",
+				description = "입력값이 유효하지 않음",
+				content = @Content(
+						mediaType = "application/json",
+						schema = @Schema(implementation = CommentInvalidInputResponse.class))),
+		@io.swagger.v3.oas.annotations.responses.ApiResponse(
+				responseCode = "401",
+				description = "로그인 되지 않은 사용자",
+				content = @Content(
+						mediaType = "application/json",
+						schema = @Schema(implementation = CommentUnauthorizedAccessResponse.class))),
+		@io.swagger.v3.oas.annotations.responses.ApiResponse(
+				responseCode = "401",
+				description = "권한 없는 사용자의 접근",
+				content = @Content(
+						mediaType = "application/json",
+						schema = @Schema(implementation = CommentForbiddenResponse.class))),
+		@io.swagger.v3.oas.annotations.responses.ApiResponse(
+				responseCode = "404",
+				description = "해당 댓글을 찾을 수 없음",
+				content = @Content(
+						mediaType = "application/json",
+						schema = @Schema(implementation = CommentNotFoundResponse.class))),
+		@io.swagger.v3.oas.annotations.responses.ApiResponse(
+				responseCode = "500",
+				description = "서버 내부 오류",
+				content = @Content(
+						mediaType = "application/json",
+						schema = @Schema(implementation = InternalServerErrorResponse.class)))
+		
+	})
 	@PostMapping("/comments")
-	public ResponseEntity<?> createComment(@RequestBody CommentCreateRequestDto commentRequestDto) {
+	public ResponseEntity<?> createComment(
+			@Parameter(description = "댓글 작성 요청 정보", required = true) @RequestBody CommentCreateRequestDto commentRequestDto) {
 		
 		// 인증 여부 확인 (비로그인)
 		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
 		if (authentication == null || !authentication.isAuthenticated() || authentication.getPrincipal().equals("anonymousUser")) {
-		    throw new ApiException(VoteErrorCode.VOTE_UNAUTHORIZED_ACCESS);
+		    throw new ApiException(CommentErrorCode.COMMENT_UNAUTHORIZED_ACCESS);
 		}
 		
 		// 권한 없는 사용자의 접근
 		String username = SecurityContextHolder.getContext().getAuthentication().getName();
 		if (userService.readUserByUsername(username).getId() != commentRequestDto.getUserId()) {
-		    throw new ApiException(VoteErrorCode.VOTE_FORBIDDEN);
+		    throw new ApiException(CommentErrorCode.COMMENT_FORBIDDEN);
 		}
 		
 		CommentCreateResponseDto commentResponseDto = commentService.createComment(commentRequestDto);
@@ -89,19 +158,70 @@ public class CommentController {
 	
 	
 	// 댓글 수정
+	@Operation(
+	    summary = "댓글 수정",
+	    description = "댓글을 수정합니다."
+	)
+	@ApiResponses(value = {
+		@io.swagger.v3.oas.annotations.responses.ApiResponse(
+				responseCode = "200",
+				description = "댓글 수정 성공",
+				content = @Content(
+						mediaType = "application/json",
+						schema = @Schema(implementation = CommentUpdateSuccessResponse.class))),
+		@io.swagger.v3.oas.annotations.responses.ApiResponse(
+				responseCode = "400",
+				description = "댓글 수정 실패",
+				content = @Content(
+						mediaType = "application/json",
+						schema = @Schema(implementation = CommentUpdateFailResponse.class))),
+		@io.swagger.v3.oas.annotations.responses.ApiResponse(
+				responseCode = "400",
+				description = "입력값이 유효하지 않음",
+				content = @Content(
+						mediaType = "application/json",
+						schema = @Schema(implementation = CommentInvalidInputResponse.class))),
+		@io.swagger.v3.oas.annotations.responses.ApiResponse(
+				responseCode = "401",
+				description = "로그인 되지 않은 사용자",
+				content = @Content(
+						mediaType = "application/json",
+						schema = @Schema(implementation = CommentUnauthorizedAccessResponse.class))),
+		@io.swagger.v3.oas.annotations.responses.ApiResponse(
+				responseCode = "401",
+				description = "권한 없는 사용자의 접근",
+				content = @Content(
+						mediaType = "application/json",
+						schema = @Schema(implementation = CommentForbiddenResponse.class))),
+		@io.swagger.v3.oas.annotations.responses.ApiResponse(
+				responseCode = "404",
+				description = "해당 댓글을 찾을 수 없음",
+				content = @Content(
+						mediaType = "application/json",
+						schema = @Schema(implementation = CommentNotFoundResponse.class))),
+		@io.swagger.v3.oas.annotations.responses.ApiResponse(
+				responseCode = "500",
+				description = "서버 내부 오류",
+				content = @Content(
+						mediaType = "application/json",
+						schema = @Schema(implementation = InternalServerErrorResponse.class)))
+		
+	})
 	@PatchMapping("/comments/{id}")
-	public ResponseEntity<?> updateComment(@PathVariable int id, @RequestBody CommentUpdateRequestDto commentRequestDto) {
+	public ResponseEntity<?> updateComment(
+			@Parameter(description = "댓글의 일련번호", required = true, example = "1") @PathVariable int id,
+			@Parameter(description = "댓글 수정 요청 정보", required = true) @RequestBody CommentUpdateRequestDto commentRequestDto) {
 		
 		// 인증 여부 확인 (비로그인)
 		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
 		if (authentication == null || !authentication.isAuthenticated() || authentication.getPrincipal().equals("anonymousUser")) {
-		    throw new ApiException(VoteErrorCode.VOTE_UNAUTHORIZED_ACCESS);
+		    throw new ApiException(CommentErrorCode.COMMENT_UNAUTHORIZED_ACCESS);
 		}
 		
 		// 권한 없는 사용자의 접근
 		String username = SecurityContextHolder.getContext().getAuthentication().getName();
 		if (userService.readUserByUsername(username).getId() != commentRequestDto.getUserId()) {
-		    throw new ApiException(VoteErrorCode.VOTE_FORBIDDEN);
+		    throw new ApiException(CommentErrorCode.COMMENT_FORBIDDEN);
 		}
 		
 		CommentUpdateResponseDto commentResponseDto = commentService.updateComment(commentRequestDto);
@@ -114,20 +234,71 @@ public class CommentController {
 	
 	// 댓글 삭제
 	@DeleteMapping("/comments/{id}")
-	public ResponseEntity<?> deleteComment(@PathVariable int id, @RequestBody CommentDeleteRequestDto commentDto) {
+	@Operation(
+	    summary = "댓글 삭제",
+	    description = "댓글을 삭제합니다."
+	)
+	@ApiResponses(value = {
+		@io.swagger.v3.oas.annotations.responses.ApiResponse(
+				responseCode = "200",
+				description = "댓글 삭제 성공",
+				content = @Content(
+						mediaType = "application/json",
+						schema = @Schema(implementation = CommentDeleteSuccessResponse.class))),
+		@io.swagger.v3.oas.annotations.responses.ApiResponse(
+				responseCode = "400",
+				description = "댓글 삭제 실패",
+				content = @Content(
+						mediaType = "application/json",
+						schema = @Schema(implementation = CommentDeleteFailResponse.class))),
+		@io.swagger.v3.oas.annotations.responses.ApiResponse(
+				responseCode = "400",
+				description = "입력값이 유효하지 않음",
+				content = @Content(
+						mediaType = "application/json",
+						schema = @Schema(implementation = CommentInvalidInputResponse.class))),
+		@io.swagger.v3.oas.annotations.responses.ApiResponse(
+				responseCode = "401",
+				description = "로그인 되지 않은 사용자",
+				content = @Content(
+						mediaType = "application/json",
+						schema = @Schema(implementation = CommentUnauthorizedAccessResponse.class))),
+		@io.swagger.v3.oas.annotations.responses.ApiResponse(
+				responseCode = "401",
+				description = "권한 없는 사용자의 접근",
+				content = @Content(
+						mediaType = "application/json",
+						schema = @Schema(implementation = CommentForbiddenResponse.class))),
+		@io.swagger.v3.oas.annotations.responses.ApiResponse(
+				responseCode = "404",
+				description = "해당 댓글을 찾을 수 없음",
+				content = @Content(
+						mediaType = "application/json",
+						schema = @Schema(implementation = CommentNotFoundResponse.class))),
+		@io.swagger.v3.oas.annotations.responses.ApiResponse(
+				responseCode = "500",
+				description = "서버 내부 오류",
+				content = @Content(
+						mediaType = "application/json",
+						schema = @Schema(implementation = InternalServerErrorResponse.class)))
+		
+	})
+	public ResponseEntity<?> deleteComment(
+			@Parameter(description = "댓글의 일련번호", required = true, example = "1") @PathVariable int id,
+			@Parameter(description = "댓글 삭제 요청 정보", required = true) @RequestBody CommentDeleteRequestDto commentDto) {
 		
 		// 인증 여부 확인 (비로그인)
-//		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-//		if (authentication == null || !authentication.isAuthenticated() || authentication.getPrincipal().equals("anonymousUser")) {
-//		    throw new ApiException(VoteErrorCode.VOTE_UNAUTHORIZED_ACCESS);
-//		}
+		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+		if (authentication == null || !authentication.isAuthenticated() || authentication.getPrincipal().equals("anonymousUser")) {
+		    throw new ApiException(CommentErrorCode.COMMENT_UNAUTHORIZED_ACCESS);
+		}
 		
 		// 권한 없는 사용자의 접근
-//		String username = SecurityContextHolder.getContext().getAuthentication().getName();
-//		if (userService.readUserByUsername(username).getId() != commentDto.getUserId()) {
-//		    throw new ApiException(VoteErrorCode.VOTE_FORBIDDEN);
-//		}
-//		
+		String username = SecurityContextHolder.getContext().getAuthentication().getName();
+		if (userService.readUserByUsername(username).getId() != commentDto.getUserId()) {
+		    throw new ApiException(CommentErrorCode.COMMENT_FORBIDDEN);
+		}
+		
 		commentService.deleteComment(commentDto);
 		
 		return ResponseEntity.status(CommentSuccessCode.COMMENT_DELETE_SUCCESS.getStatus())
